Updating python version in tutorial
We need to update the version of the python module in the docs, since starting in v4 of GenPipes, we need to use python 3 instead of python 2 (as is currently indicated in this page). Maybe we can add a note to indicate that you need a different version of python depending on the version of GenPipes.